+++ drivers/csv/detect_header.go
@@ -4,10 +4,11 @@ import (
 	"context"
 	"strings"
 
+	"github.com/neilotoole/sq/libsq/driver"
+
 	"github.com/neilotoole/sq/libsq/core/lg"
 	"github.com/neilotoole/sq/libsq/core/lg/lga"
 
-	"github.com/neilotoole/sq/drivers"
 	"github.com/neilotoole/sq/libsq/core/options"
 
 	"github.com/neilotoole/sq/libsq/core/errz"
@@ -18,10 +19,10 @@ import (
 // set in opts, or if detectHeaderRow detects that the first
 // row of recs seems to be a header.
 func hasHeaderRow(ctx context.Context, recs [][]string, opts options.Options) (bool, error) {
-	if drivers.OptIngestHeader.IsSet(opts) {
-		b := drivers.OptIngestHeader.Get(opts)
+	if driver.OptIngestHeader.IsSet(opts) {
+		b := driver.OptIngestHeader.Get(opts)
 		lg.FromContext(ctx).Debug("CSV ingest header explicitly specified: skipping header detection",
-			lga.Key, drivers.OptIngestHeader.Key(),
+			lga.Key, driver.OptIngestHeader.Key(),
 			lga.Val, b)
 		return b, nil
 	}