+++ drivers/xlsx/ingest.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/neilotoole/sq/drivers"
 	"github.com/neilotoole/sq/libsq/core/lg/lga"
 
 	"github.com/neilotoole/sq/libsq/core/lg/lgm"
@@ -35,7 +34,7 @@ func xlsxToScratch(ctx context.Context, src *source.Source, xlFile *xlsx.File, s
 		lga.Src, src,
 		lga.Target, scratchDB.Source())
 
-	hasHeader := drivers.OptIngestHeader.Get(src.Options)
+	hasHeader := driver.OptIngestHeader.Get(src.Options)
 
 	// TODO: Like the csv driver, the xlsx driver should detect
 	// the presence of a header.
@@ -189,7 +188,7 @@ func buildTblDefsForSheets(ctx context.Context, sheets []*xlsx.Sheet, hasHeader 
 			default:
 			}
 
-			tblDef, err := buildTblDefForSheet(lg.FromContext(gCtx), sheets[i], hasHeader)
+			tblDef, err := buildTblDefForSheet(gCtx, sheets[i], hasHeader)
 			if err != nil {
 				return err
 			}
@@ -208,7 +207,8 @@ func buildTblDefsForSheets(ctx context.Context, sheets []*xlsx.Sheet, hasHeader 
 // buildTblDefForSheet creates a table for the given sheet, and returns
 // a model of the table, or an error. If the sheet is empty, (nil,nil)
 // is returned.
-func buildTblDefForSheet(log *slog.Logger, sheet *xlsx.Sheet, hasHeader bool) (*sqlmodel.TableDef, error) {
+func buildTblDefForSheet(ctx context.Context, sheet *xlsx.Sheet, hasHeader bool) (*sqlmodel.TableDef, error) {
+	log := lg.FromContext(ctx)
 	maxCols := getRowsMaxCellCount(sheet)
 	if maxCols == 0 {
 		log.Warn("sheet is empty: skipping", "sheet", sheet.Name)
@@ -262,6 +262,11 @@ func buildTblDefForSheet(log *slog.Logger, sheet *xlsx.Sheet, hasHeader bool) (*
 	}
 
 	colNames, colKinds = syncColNamesKinds(colNames, colKinds)
+
+	var err error
+	if colNames, err = driver.MungeIngestColNames(ctx, colNames); err != nil {
+		return nil, err
+	}
 
 	tblDef := &sqlmodel.TableDef{Name: sheet.Name}
 	cols := make([]*sqlmodel.ColDef, len(colNames))