+++ drivers/xlsx/xlsx.go
@@ -6,7 +6,6 @@ import (
 	"database/sql"
 	"io"
 
-	"github.com/neilotoole/sq/drivers"
 	"github.com/neilotoole/sq/libsq/core/lg/lga"
 
 	"github.com/neilotoole/sq/libsq/core/lg/lgm"
@@ -231,7 +230,7 @@ func (d *database) SourceMetadata(_ context.Context, noSchema bool) (*source.Met
 		return meta, nil
 	}
 
-	hasHeader := drivers.OptIngestHeader.Get(d.src.Options)
+	hasHeader := driver.OptIngestHeader.Get(d.src.Options)
 	for _, sheet := range xlFile.Sheets {
 		tbl := &source.TableMetadata{Name: sheet.Name, RowCount: int64(len(sheet.Rows))}
 
@@ -273,7 +272,7 @@ func (d *database) TableMetadata(_ context.Context, tblName string) (*source.Tab
 		return nil, errz.Wrapf(err, "unable to open XLSX file: %s", d.src.Location)
 	}
 
-	hasHeader := drivers.OptIngestHeader.Get(d.src.Options)
+	hasHeader := driver.OptIngestHeader.Get(d.src.Options)
 
 	for _, sheet := range xlFile.Sheets {
 		if sheet.Name != tblName {